Moved the User related routes into their own file instead of index. Also narrowed down the user profile page into one page.
