+++ views/job/application/message.pug
@@ -23,7 +23,7 @@ block content
                                 | on the description of the position, what do you think makes you a valuable candidate and what previous experience puts you in a good position to handle the responsibilities of this job?
                             .row
                                 .col-lg-6.col-sm-6.mb-2
-                                    a.btn.btn-outline-primary.btn-block(href='/user/teacher-details/' + user.id )
+                                    a.btn.btn-outline-primary.btn-block(href='/user/' )
                                         i.mdi.mdi-arrow-left.mr-2(aria-hidden='true')
                                         | View Teacher Profile
                                 .col-lg-6.col-sm-6.mb-2
