--- views/login/edit-user.pug
@@ -1,6 +1,0 @@
-//- .border.border-pale-gray.border-radius-5.bg-white.text-dark.px-3.py-1.px-md-5.mb-3   
-//-     .row.align-items-center
-//-         .col-lg-2.text-right.mb-3
-//-             img.img-prompt.img-prompt-small.img-fluid(src='https://englishinchina.s3.amazonaws.com/lg_be9d0ab0-1de5-11e9-9ef8-4fcb4d192528.jpeg')
-//-         #write-message-prompt.col-lg-10
-//-             h1.text-italic @#{user.username} - #{user.email}
