--- views/login/edituser.pug
@@ -1,51 +1,0 @@
-extends form-container
-block card
-    .col-md-10.col-sm-12.col-lg-9
-        .border.border-pale-gray.border-radius-5.text-dark.bg-white
-            a.collapse-link(data-toggle='collapse', href='#teacher-profile-collapse', role='button', aria-expanded='false', aria-controls='teacher-profile-collapse')
-                h1.text-muted.my-5.font-weight-bold 
-                    | Teacher Profile      
-                    i#teacher-profile-collapse-chevron.mdi.mdi-chevron-up(aria-hidden='true')
-        .collapse.show#teacher-profile-collapse     
-
-            .my-5.text-left.px-3.px-md-5.py-1
-                h3.mb-5.text-dark.font-weight-bold My Profile
-                form(class='form-signin', action='/user/edit', method='POST')
-
-                    if(redirectUrl)
-                        .form-group.d-none
-                                input(type='text', name='redirectUrl', placeholder='redirect', value=redirectUrl)
-
-                    .form-group
-                        input#id(type="hidden", name="id", value=user.id)
-
-                    .row
-                        .col-12
-                            -var textInput = {}
-                            -textInput.id = 'username'
-                            -textInput.name = 'username'
-                            -textInput.placeholder = 'Username'
-                            -textInput.label = 'Username'
-                            -textInput.required = true;
-                            include ../templates/inputs/text-input
-
-                    .row
-                        .col-12
-                            -var textInput = {}
-                            -textInput.id = 'password'
-                            -textInput.name = 'password'
-                            -textInput.type = 'password'
-                            -textInput.placeholder = '********'
-                            -textInput.label = 'Password'
-                            -textInput.required = true;
-                            -textInput.disabled = true;
-                            include ../templates/inputs/text-input
-                            p.text-small
-                                a(href="/forgot") Request password change
-
-                    .row
-                        .col-12.text-center.mt-5
-                            button.btn.btn-md.btn-block.btn-primary.btn-lg.px-5.px-5.mb-1(type='submit') Save Changes
-
-block dashboard
-    include user-dashboard
