+++ views/login/my-account.pug
@@ -1,11 +1,12 @@
 .border.border-pale-gray.border-radius-5.bg-white.text-dark.px-3.px-md-5
     a.collapse-link(data-toggle='collapse', href='#my-account-collapse', role='button', aria-expanded='false', aria-controls='my-account-collapse')
         h1.text-muted.my-5.font-weight-bold 
-            | My Account      
+            i.mdi.mdi-account-box-outline(aria-hidden='true')
+            |  My Account      
             i#my-account-collapse-chevron.mdi.mdi-chevron-down(aria-hidden='true')
     .collapse#my-account-collapse  
         h3.mb-3.text-dark.font-weight-bold My Profile
-        form(class='form-signin', action='/user/edit', method='POST')
+        form(class='form-signin', action='/user', method='POST')
 
             if(redirectUrl)
                 .form-group.d-none
@@ -27,6 +28,17 @@
 
                 .col-lg-6
                     -var textInput = {}
+                    -textInput.id = 'email'
+                    -textInput.name = 'email'
+                    -textInput.placeholder = 'Email'
+                    -textInput.label = 'Email'
+                    -textInput.required = true;
+                    -textInput.value = user.email
+                    include ../templates/inputs/text-input
+
+            .row
+                .col-lg-6
+                    -var textInput = {}
                     -textInput.id = 'password'
                     -textInput.name = 'password'
                     -textInput.type = 'password'
@@ -34,19 +46,7 @@
                     -textInput.label = 'Password'
                     -textInput.required = true;
                     -textInput.disabled = true;
-                    include ../templates/inputs/text-input
-                    p.text-small
-                        a(href="/forgot") Request password change
-
-            .row
-                .col-lg-6
-                    -var textInput = {}
-                    -textInput.id = 'email'
-                    -textInput.name = 'email'
-                    -textInput.placeholder = 'Email'
-                    -textInput.label = 'Email'
-                    -textInput.required = true;
-                    -textInput.value = user.email
+                    -textInput.undertext = {value:'Request password change', link:'/forgot'}
                     include ../templates/inputs/text-input
 
             .row
